+++ src/main/java/isw21/descuentos/OfertaFactory.java
@@ -2,7 +2,7 @@ package main.java.isw21.descuentos;
 
 public class OfertaFactory extends AbstractFactory{
     @Override
-    public Oferta getOferta(String txtEntidadText, String fechaIni, String fechaFin, int i, String text, int p, int gastado) {
+    public Oferta getOferta(String txtEntidadText, String fechaIni, String fechaFin, int i, String text, int p, double gastado) {
         if (p == 0){
             return new Descuento(txtEntidadText,fechaIni,fechaFin,i,text);
         }else if(p ==1){
